Move time-related aliases to utils.predictions
These types are often used together with other parts of
`cmk.utils.predictions`.  This change therefore improves
the cohesion.  And when they're *not* used with other
types or functions of `predictions`, then it's most likely
the wrong kind of coupling and it should be fixed.

These types are just aliases to int anyway so they are
not really useful.

CMK-13768
