+++ cmk/ec/core_queries.py
@@ -9,8 +9,9 @@ from typing import Any, NamedTuple
 from livestatus import LocalConnection
 
 from cmk.utils.hostaddress import HostAddress, HostName
+from cmk.utils.prediction import Timestamp
 from cmk.utils.timeperiod import TimeperiodName
-from cmk.utils.type_defs import ContactgroupName, Timestamp
+from cmk.utils.type_defs import ContactgroupName
 from cmk.utils.user import UserId
