+++ cmk/gui/painter/v0/painters.py
@@ -14,8 +14,8 @@ import cmk.utils.paths
 import cmk.utils.version as cmk_version
 from cmk.utils.defines import short_host_state_name, short_service_state_name
 from cmk.utils.labels import Labels
+from cmk.utils.prediction import Timestamp
 from cmk.utils.render import approx_age
-from cmk.utils.type_defs import Timestamp
 
 import cmk.gui.metrics as metrics
 import cmk.gui.sites as sites
