+++ cmk/gui/plugins/metrics/graph_images.py
@@ -16,7 +16,7 @@ import livestatus
 
 from cmk.utils.exceptions import MKGeneralException
 from cmk.utils.hostaddress import HostName
-from cmk.utils.type_defs import Timestamp
+from cmk.utils.prediction import Timestamp
 
 import cmk.gui.pdf as pdf
 from cmk.gui.config import active_config
