+++ cmk/gui/valuespec.py
@@ -72,8 +72,8 @@ import cmk.utils.regex
 from cmk.utils.encryption import Encrypter, fetch_certificate_details
 from cmk.utils.exceptions import MKGeneralException
 from cmk.utils.plugin_registry import Registry
+from cmk.utils.prediction import Seconds, TimeRange
 from cmk.utils.render import SecondsRenderer
-from cmk.utils.type_defs import Seconds, TimeRange
 from cmk.utils.urls import is_allowed_url
 from cmk.utils.user import UserId
 from cmk.utils.version import Version
